Rename stub test artifacts for consistency
